# openapi.yaml
openapi: 3.0.3
info:
  title: XMiete Core API
  description: Open-source standard for digital rental deposits (BGB § 551 compliant).
  version: 1.0.0
  contact:
    name: XMiete Open Source
    url: https://xmiete.org
servers:
  - url: https://api.xmiete.org/v1
    description: Production API
paths:
  /deposits:
    post:
      summary: Create a new deposit request
      operationId: createDeposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './xmiete_schema.json'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: './xmiete_schema.json'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT signed with JWT_SECRET — used for deposit lifecycle endpoints.
    CredentialBearerAuth:
      type: http
      scheme: bearer
      description: Short-lived access token issued by POST /v1/token — used for POST /v1/credential.
security:
  - BearerAuth: []